refactor(ui): move shared ui to a separate library
moved shared-core ui directory to a separate library;
confirmed that eventhing works: linting,
testing, starting
