+++ libs/shared-store/src/lib/state/http-api/http-api.service.spec.ts
@@ -7,7 +7,7 @@ import {
   newTestBedMetadata,
 } from '@nx-ng-starter/mocks-core';
 import { HttpHandlersService, ToasterService } from '@nx-ng-starter/shared-core/data-access';
-import { AppTranslateModule, CustomMaterialModule } from '@nx-ng-starter/shared-core/ui';
+import { AppTranslateModule, CustomMaterialModule } from '@nx-ng-starter/shared-ui';
 import { Apollo, ApolloModule } from 'apollo-angular';
 import { HttpLinkModule } from 'apollo-angular-link-http';
 import { of } from 'rxjs';
