+++ libs/shared-ui/src/lib/modules/custom-material/hammerjs-gesture.config.ts
