+++ nx.json
@@ -35,6 +35,9 @@
     },
     "shared-util": {
       "tags": ["scope:nx-ng-starter", "type:util", "platform:desktop"]
+    },
+    "shared-ui": {
+      "tags": ["scope:nx-ng-starter", "type:ui", "platform:desktop"]
     }
   },
   "tasksRunnerOptions": {
