+++ tools/shell/module-aliases.sh
@@ -21,6 +21,7 @@ MODULE_ALIAS_LIB_MOCKS_CORE="lib:mocks-core"
 MODULE_ALIAS_LIB_SHARED_CORE="lib:shared-core"
 MODULE_ALIAS_LIB_SHARED_STORE="lib:shared-store"
 MODULE_ALIAS_LIB_SHARED_UTIL="lib:shared-util"
+MODULE_ALIAS_LIB_SHARED_UI="lib:shared-ui"
 MODULE_ALIAS_LIB_PROTO="lib:proto"
 
 MODULE_ALIAS_NAMES=(
@@ -32,6 +33,7 @@ MODULE_ALIAS_NAMES=(
   MODULE_ALIAS_LIB_SHARED_CORE
   MODULE_ALIAS_LIB_SHARED_STORE
   MODULE_ALIAS_LIB_SHARED_UTIL
+  MODULE_ALIAS_LIB_SHARED_UI
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -44,6 +46,7 @@ MODULE_ALIAS_VARS=(
   "$MODULE_ALIAS_LIB_SHARED_CORE"
   "$MODULE_ALIAS_LIB_SHARED_STORE"
   "$MODULE_ALIAS_LIB_SHARED_UTIL"
+  "$MODULE_ALIAS_LIB_SHARED_UI"
   "$MODULE_ALIAS_LIB_PROTO"
 )
 
@@ -55,6 +58,7 @@ MODULE_ALIAS_NAMES_UNIT=(
   MODULE_ALIAS_LIB_SHARED_CORE
   MODULE_ALIAS_LIB_SHARED_STORE
   MODULE_ALIAS_LIB_SHARED_UTIL
+  MODULE_ALIAS_LIB_SHARED_UI
   MODULE_ALIAS_LIB_PROTO
 )
 
@@ -66,6 +70,7 @@ MODULE_ALIAS_VARS_UNIT=(
   "$MODULE_ALIAS_LIB_SHARED_CORE"
   "$MODULE_ALIAS_LIB_SHARED_STORE"
   "$MODULE_ALIAS_LIB_SHARED_UTIL"
+  "$MODULE_ALIAS_LIB_SHARED_UI"
   "$MODULE_ALIAS_LIB_PROTO"
 )
