+++ tsconfig.json
@@ -51,6 +51,12 @@
     },
     {
       "path": "./libs/shared-util/tsconfig.spec.json"
+    },
+    {
+      "path": "./libs/shared-ui/tsconfig.lib.json"
+    },
+    {
+      "path": "./libs/shared-ui/tsconfig.spec.json"
     }
   ]
 }
